Changed image storage strategy

+++ src/Demergenza/Controllers/CategoryController.cs
@@ -1,6 +1,7 @@
 using Demergenza.Application.Abstractions.Repositories.AdminRepository;
 using Demergenza.Application.Abstractions.Repositories.CategoryRepository;
 using Demergenza.Application.DTOs.Category;
+using Demergenza.Application.DTOs.Menu;
 using Demergenza.Application.Services;
 using Demergenza.Domain.Entities.Admin;
 using Demergenza.Domain.Entities.Menu;
@@ -32,37 +33,60 @@ namespace Demergenza.Controllers
         [Route("getcategorybyid/{id}")]
         public async Task<IActionResult> GetCategoryById([FromRoute] string id)
         {
-            Category? category = await _categoryRead.GetByIdAsync(Guid.Parse(id));
-            return Ok(category);
+            Category category = await _categoryRead.GetByIdAsync(Guid.Parse(id));
+            return Ok(new CategoryResponse()
+            {
+                Id = category.Id,
+                Admin = category.Admin,
+                Image = category.Image,
+                Menus = category.Menus.Select(m => new MenuResponse()
+                {
+                    Name = m.Name,
+                    Image = m.Name,
+                    Date = m.Date,
+                    Ingredients = m.Ingredients,
+                    Price = m.Price,
+                    Id = m.Id
+                }).ToList(),
+                Name = category.Name
+            });
         }
 
         [HttpGet]
         [Route("getallcategories")]
         public IActionResult GetAllCategories()
         {
-            ICollection<Category> categories = _categoryRead.Select(c => new Category()
+            IQueryable<CategoryResponse> categories = _categoryRead.GetAll().Select(c => new CategoryResponse()
             {
                 Id = c.Id,
-                Menus = c.Menus,
-                Name = c.Name,
-                Image = c.Image
-            }).ToList();
+                Image = c.Image,
+                Menus = c.Menus.Select(m => new MenuResponse()
+                {
+                    Name = m.Name,
+                    Image = m.Name,
+                    Date = m.Date,
+                    Ingredients = m.Ingredients,
+                    Price = m.Price,
+                    Id = m.Id
+                }).ToList(),
+                Name = c.Name
+            });
             return Ok(categories);
         }
 
         [Authorize]
         [HttpPost]
         [Route("addcategory")]
-        public async Task<IActionResult> AddCategory([FromForm] AddCategory addacategory)
+        public async Task<IActionResult> AddCategory([FromForm] AddCategory addCategory)
         {
-            string newImageName = _imageService.SaveImage(addacategory.categoryImage);
-            Admin? admin = await _adminRead.GetFirstAsync(admin => admin.Username == addacategory.AdminUsername);
+            string newImageName = _imageService.SaveImage(addCategory.categoryImage);
+            Admin? admin = await _adminRead.GetFirstAsync(admin => admin.Username == addCategory.AdminUsername);
             if (admin is null) return Unauthorized();
             Category category = new()
             {
                 Id = Guid.NewGuid(),
-                Name = addacategory.CategoryName,
-                Image = $"{Request.Scheme}://{Request.Host}/data-images/{newImageName}",
+                Name = addCategory.CategoryName,
+                Image = newImageName,
                 Admin = admin,
                 Date = DateTime.UtcNow
             };
@@ -86,8 +110,10 @@ namespace Demergenza.Controllers
 
             await _categoryWrite.RemoveAsync(Guid.Parse(id));
             await _categoryWrite.SaveAsync();
-            bool isDeleted = _imageService.DeleteImageByName(Path.GetFileName(category.Image));
-            return Ok(isDeleted);
+            if (category.Image is not null)
+                _imageService.DeleteImageByName(Path.GetFileName(category.Image));
+
+            return Ok();
         }
 
         [Authorize]
@@ -104,10 +130,12 @@ namespace Demergenza.Controllers
             {
                 string? oldImageName = Path.GetFileName(category.Image);
                 string newImageName = _imageService.SaveImage(categoryModel.CategoryImage);
-                category.Image = $"{Request.Scheme}://{Request.Host}/data-images/{newImageName}";
-                _imageService.DeleteImageByName(oldImageName);
+                category.Image = newImageName;
+                if (oldImageName is not null)
+                {
+                    _imageService.DeleteImageByName(oldImageName);
+                }
             }
-
 
             Admin? admin = await _adminRead.GetFirstAsync(admin => admin.Username == categoryModel.AdminUsername);
             category.Admin = admin ?? category.Admin;
